Add files via upload
This model has a better accuracy. The accuracy can be improved further by trying out different linear models in place of SGD and also by varying the training testing data. One word of advise would be to drop the empty values rather than filling it with values. Also do check for outliers, i.e values that do not correspond to the trend of rest of the data. You can also try using other ensembles than HistGradientBoostingRegressor, also you can try varying the max_depths but I did it over a good range and the optimal value is in the code.